requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
anyio==4.2.0
attrs==23.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
blinker==1.8.2
build==1.2.1
cachecontrol==0.14.0
cachetools==5.3.3
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
colorama==0.4.6
crashtest==0.4.1
crewai==0.11.0
dataclasses-json==0.6.4
deepdiff==7.0.1
deprecated==1.2.14
distlib==0.3.8
distro==1.9.0
docstring-parser==0.15
dulwich==0.21.7
emoji==2.12.1
fastjsonschema==2.19.1
filelock==3.14.0
filetype==1.2.0
frozenlist==1.4.1
gitdb==4.0.11
gitpython==3.1.43
google-search-results==2.4.2
googleapis-common-protos==1.62.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
importlib-metadata==6.11.0
installer==0.7.0
instructor==0.5.2
jaraco-classes==3.4.0
jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
keyring==24.3.1
langchain==0.1.7
langchain-community==0.0.20
langchain-core==0.1.23
langchain-openai==0.0.2.post1
langdetect==1.0.9
langsmith==0.0.87
lxml==5.2.2
markdown-it-py==3.0.0
markupsafe==2.1.5
marshmallow==3.20.2
mdurl==0.1.2
more-itertools==10.2.0
msgpack==1.0.8
multidict==6.0.5
mypy-extensions==1.0.0
nltk==3.8.1
numpy==1.26.4
openai==1.12.0
opencv-python==4.8.1.78
opentelemetry-api==1.22.0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-http==1.22.0
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
ordered-set==4.1.0
packaging==23.2
pandas==2.2.2
pexpect==4.9.0
pillow==10.3.0
pip==24.0
pkginfo==1.10.0
platformdirs==4.2.2
poetry==1.8.3
poetry-core==1.9.0
poetry-plugin-export==1.8.0
protobuf==4.25.2
ptyprocess==0.7.0
pyarrow==16.1.0
pydantic==2.6.1
pydantic-core==2.16.2
pydeck==0.9.1
pygments==2.17.2
pypdf==4.2.0
pyproject-hooks==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-iso639==2024.4.27
python-magic==0.4.27
pytz==2024.1
pywin32-ctypes==0.2.2
pyyaml==6.0.1
rapidfuzz==3.9.1
referencing==0.35.1
regex==2023.12.25
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.0
rpds-py==0.18.1
setuptools==69.5.1
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
sqlalchemy==2.0.27
streamlit==1.34.0
tabulate==0.9.0
tenacity==8.2.3
tiktoken==0.5.2
toml==0.10.2
tomlkit==0.12.5
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
trove-classifiers==2024.5.17
typer==0.9.0
typing-extensions==4.9.0
typing-inspect==0.9.0
tzdata==2024.1
unstructured==0.14.0
unstructured-client==0.22.0
urllib3==2.2.0
virtualenv==20.26.2
watchdog==4.0.0
wheel==0.43.0
wrapt==1.16.0
yarl==1.9.4
zipp==3.17.0